+++ client/src/components/Card/MyCardInfo.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const MyCardInfo = () => {
+  return (
+    <div>
+      {/* To-do: 내 카드 데이터가 없으면 NoCard 컴포넌트 띄우기!! */}
+      my card
+    </div>
+  );
+};
+
+export default MyCardInfo;
